Moved file into Helper
